tor: updated to 0.2.9.8

+++ release/src/router/tor/src/test/test_channeltls.c
@@ -1,6 +1,8 @@
 /* Copyright (c) 2014-2016, The Tor Project, Inc. */
 /* See LICENSE for licensing information */
 
+#include "orconfig.h"
+
 #include <math.h>
 
 #define TOR_CHANNEL_INTERNAL_
@@ -124,8 +126,7 @@ test_channeltls_num_bytes_queued(void *arg)
    * Next, we have to test ch->num_bytes_queued, which is
    * channel_tls_num_bytes_queued_method.  We can't mock
    * connection_get_outbuf_len() directly because it's static inline
-   * in connection.h, but we can mock buf_datalen().  Note that
-   * if bufferevents ever work, this will break with them enabled.
+   * in connection.h, but we can mock buf_datalen().
    */
 
   tt_assert(ch->num_bytes_queued != NULL);
@@ -185,7 +186,7 @@ test_channeltls_overhead_estimate(void *arg)
   const char test_digest[DIGEST_LEN] = {
     0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a,
     0x0b, 0x0c, 0x0d, 0x0e, 0x0f, 0x10, 0x11, 0x12, 0x13, 0x14 };
-  float r;
+  double r;
   channel_tls_t *tlschan = NULL;
 
   (void)arg;
@@ -206,31 +207,31 @@ test_channeltls_overhead_estimate(void *arg)
   ch = channel_tls_connect(&test_addr, 567, test_digest);
   tt_assert(ch != NULL);
 
-  /* First case: silly low ratios should get clamped to 1.0f */
+  /* First case: silly low ratios should get clamped to 1.0 */
   tlschan = BASE_CHAN_TO_TLS(ch);
   tt_assert(tlschan != NULL);
   tlschan->conn->bytes_xmitted = 128;
   tlschan->conn->bytes_xmitted_by_tls = 64;
   r = ch->get_overhead_estimate(ch);
-  tt_assert(fabsf(r - 1.0f) < 1E-12);
+  tt_assert(fabs(r - 1.0) < 1E-12);
 
   tlschan->conn->bytes_xmitted_by_tls = 127;
   r = ch->get_overhead_estimate(ch);
-  tt_assert(fabsf(r - 1.0f) < 1E-12);
+  tt_assert(fabs(r - 1.0) < 1E-12);
 
   /* Now middle of the range */
   tlschan->conn->bytes_xmitted_by_tls = 192;
   r = ch->get_overhead_estimate(ch);
-  tt_assert(fabsf(r - 1.5f) < 1E-12);
+  tt_assert(fabs(r - 1.5) < 1E-12);
 
-  /* Now above the 2.0f clamp */
+  /* Now above the 2.0 clamp */
   tlschan->conn->bytes_xmitted_by_tls = 257;
   r = ch->get_overhead_estimate(ch);
-  tt_assert(fabsf(r - 2.0f) < 1E-12);
+  tt_assert(fabs(r - 2.0) < 1E-12);
 
   tlschan->conn->bytes_xmitted_by_tls = 512;
   r = ch->get_overhead_estimate(ch);
-  tt_assert(fabsf(r - 2.0f) < 1E-12);
+  tt_assert(fabs(r - 2.0) < 1E-12);
 
  done:
   if (ch) {
